+++ src/main/java/com/hlysine/create_connected/CreateConnected.java
@@ -64,7 +64,6 @@ public class CreateConnected {
         CCConfigs.register(ModLoadingContext.get());
         CCConfigs.common().register();
 
-        CCItemAttributes.register();
         CCInteractionBehaviours.register();
         CCMovementBehaviours.register();