+++ src/main/java/com/hlysine/create_connected/config/CServer.java
@@ -11,6 +11,12 @@ public class CServer extends ConfigBase {
     private final ConfigBool applicationRemainingItemFix = b(true, "applicationRemainingItemFix", Comments.applicationRemainingItemFix);
     public static final Supplier<Boolean> ApplicationRemainingItemFix = CCConfigs.safeGetter(() -> CCConfigs.server().applicationRemainingItemFix.get(), true);
 
+    private final ConfigFloat vesselBoilerStressMultiplier = f(0.8125f, 0.0f, "vesselBoilerStressMultiplier", Comments.vesselBoilerStressMultiplier);
+    public static final Supplier<Double> VesselBoilerStressMultiplier = CCConfigs.safeGetter(() -> CCConfigs.server().vesselBoilerStressMultiplier.get(), 0.8);
+
+    private final ConfigFloat vesselHeatMultiplier = f(0.5f, 0.0f, "vesselHeatMultiplier", Comments.vesselHeatMultiplier);
+    public static final Supplier<Double> VesselHeatMultiplier = CCConfigs.safeGetter(() -> CCConfigs.server().vesselHeatMultiplier.get(), 0.5);
+
     public final CStress stressValues = nested(0, CStress::new, Comments.stress);
     public final ConfigFloat brakeActiveStress = f(16384, Float.NEGATIVE_INFINITY, Float.POSITIVE_INFINITY, "brakeActiveStress", Comments.brakeActiveStress);
 
@@ -25,5 +31,7 @@ public class CServer extends ConfigBase {
         static String brakeActiveStress = "Stress impact of a powered brake [in Stress Units]";
         static String schematicsNestingDepth = "Number of sub-folders allowed for schematics";
         static String applicationRemainingItemFix = "Enable the fix that gives remaining items after manual application";
+        static String vesselBoilerStressMultiplier = "Modify the stress output when a Fluid Vessel is used as a boiler";
+        static String vesselHeatMultiplier = "Modify the heat level when a Fluid Vessel is used as a boiler";
     }
 }