+++ src/main/java/com/hlysine/create_connected/content/crankwheel/CrankWheelBlockEntity.java
@@ -1,0 +1,59 @@
+package com.hlysine.create_connected.content.crankwheel;
+
+import com.hlysine.create_connected.CCPartialModels;
+import com.jozufozu.flywheel.api.Instancer;
+import com.jozufozu.flywheel.api.Material;
+import com.jozufozu.flywheel.core.materials.model.ModelData;
+import com.simibubi.create.content.kinetics.base.IRotate;
+import com.simibubi.create.content.kinetics.crank.HandCrankBlock;
+import com.simibubi.create.content.kinetics.crank.HandCrankBlockEntity;
+import com.simibubi.create.content.kinetics.simpleRelays.ICogWheel;
+import com.simibubi.create.foundation.render.CachedBufferer;
+import com.simibubi.create.foundation.render.SuperByteBuffer;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+
+import java.util.List;
+
+public class CrankWheelBlockEntity extends HandCrankBlockEntity {
+    public CrankWheelBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
+        super(type, pos, state);
+    }
+
+    @Override
+    public List<BlockPos> addPropagationLocations(IRotate block, BlockState state, List<BlockPos> neighbours) {
+        if (!ICogWheel.isLargeCog(state))
+            return super.addPropagationLocations(block, state, neighbours);
+
+        BlockPos.betweenClosedStream(new BlockPos(-1, -1, -1), new BlockPos(1, 1, 1))
+                .forEach(offset -> {
+                    if (offset.distSqr(BlockPos.ZERO) == 2)
+                        neighbours.add(worldPosition.offset(offset));
+                });
+        return neighbours;
+    }
+
+    @Override
+    @OnlyIn(Dist.CLIENT)
+    public SuperByteBuffer getRenderedHandle() {
+        BlockState blockState = getBlockState();
+        Direction facing = blockState.getOptionalValue(HandCrankBlock.FACING)
+                .orElse(Direction.UP);
+        boolean isLarge = blockState.getBlock() instanceof CrankWheelBlock block && block.largeCog;
+        return CachedBufferer.partialFacing(isLarge ? CCPartialModels.LARGE_CRANK_WHEEL_HANDLE : CCPartialModels.CRANK_WHEEL_HANDLE, blockState, facing.getOpposite());
+    }
+
+    @Override
+    @OnlyIn(Dist.CLIENT)
+    public Instancer<ModelData> getRenderedHandleInstance(Material<ModelData> material) {
+        BlockState blockState = getBlockState();
+        Direction facing = blockState.getOptionalValue(HandCrankBlock.FACING)
+                .orElse(Direction.UP);
+        boolean isLarge = blockState.getBlock() instanceof CrankWheelBlock block && block.largeCog;
+        return material.getModel(isLarge ? CCPartialModels.LARGE_CRANK_WHEEL_HANDLE : CCPartialModels.CRANK_WHEEL_HANDLE, blockState, facing.getOpposite());
+    }
+}