+++ src/main/java/com/hlysine/create_connected/content/fluidvessel/CreativeFluidVesselBlockEntity.java
@@ -1,0 +1,28 @@
+package com.hlysine.create_connected.content.fluidvessel;
+
+import com.simibubi.create.content.fluids.tank.CreativeFluidTankBlockEntity;
+import com.simibubi.create.foundation.fluid.SmartFluidTank;
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+
+import java.util.List;
+
+public class CreativeFluidVesselBlockEntity extends FluidVesselBlockEntity {
+
+	public CreativeFluidVesselBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
+		super(type, pos, state);
+	}
+
+	@Override
+	protected SmartFluidTank createInventory() {
+		return new CreativeFluidTankBlockEntity.CreativeSmartFluidTank(getCapacityMultiplier(), this::onFluidStackChanged);
+	}
+
+	@Override
+	public boolean addToGoggleTooltip(List<Component> tooltip, boolean isPlayerSneaking) {
+		return false;
+	}
+
+}