+++ src/main/java/com/hlysine/create_connected/content/linkedtransmitter/LinkedLeverBlock.java
@@ -92,6 +92,9 @@ public class LinkedLeverBlock extends LeverBlock implements IBE<LinkedTransmitte
                                           @NotNull Player player,
                                           @NotNull InteractionHand hand,
                                           @NotNull BlockHitResult hit) {
+        if (player.isSpectator())
+            return InteractionResult.PASS;
+
         if (isHittingBase(state, level, pos, hit)) {
             if (!player.isShiftKeyDown())
                 return super.use(state, level, pos, player, hand, hit);