+++ src/main/java/com/hlysine/create_connected/datagen/CCTagGen.java
@@ -7,12 +7,11 @@ import com.hlysine.create_connected.compat.Mods;
 import com.simibubi.create.AllTags;
 import com.tterrag.registrate.providers.ProviderType;
 import com.tterrag.registrate.providers.RegistrateTagsProvider;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
-import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
-import net.minecraftforge.registries.ForgeRegistries;
 
 public class CCTagGen {
     public static void addGenerators() {
@@ -21,9 +20,13 @@ public class CCTagGen {
     }
 
     private static void genBlockTags(RegistrateTagsProvider<Block> prov) {
-        prov.tag(TagKey.create(ForgeRegistries.Keys.BLOCKS, Mods.DIAGONAL_FENCES.rl("non_diagonal_fences")))
+        prov.tag(BlockTags.create(Mods.DIAGONAL_FENCES.rl("non_diagonal_fences")))
                 .add(CCBlocks.COPYCAT_FENCE.get())
                 .add(CCBlocks.WRAPPED_COPYCAT_FENCE.get());
+        prov.tag(BlockTags.create(Mods.DREAMS_DESIRES.rl("fan_processing_catalysts_freezing")))
+                .add(CCBlocks.FAN_FREEZING_CATALYST.get());
+        prov.tag(BlockTags.create(Mods.GARNISHED.rl("fan_processing_catalysts/freezing")))
+                .add(CCBlocks.FAN_FREEZING_CATALYST.get());
     }
 
     private static void genItemTags(RegistrateTagsProvider<Item> prov) {